Convert deployment scripts to use Hardhat with zkSync Era plugins
For proper compatability with EraVM, we add and configure Hardhat to the repo with
the zkSync plugins, and use them to write deploy scripts. We keep Foundry for testing.

Note this commit also downgrades OpenZeppelin contracts to use OZ 4.x rather than OZ
5.x, for compatibility with the zksync upgradable plugin. The OpenZeppelin Foundry
upgrades plugin is completely removed, as are any Foundry tests relating to upgrades.